docs(checkbox): improve it
